Add redux action creator
To every action there must be an action creator. The action creator tells what should happen when an action takes place. For example, if a delete action is initiated, it's the action delete action creator that ensures the deleted item is actually deleted. The changes made to this document spells out what each action should really do.